requirements.txt
requests_mock==1.8.0
pytest==6.1.2
click==7.1.2
requests==2.20.0
tabulate==0.8.7
pandas==0.24.2
psycopg2_binary==2.8.6
pytest-cov==2.10.1
